+++ packages/tour/vite.config.ts
@@ -1,12 +1,6 @@
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
-import { resolve } from 'path';
 
 export default defineConfig({
-	plugins: [sveltekit()],
-	resolve: {
-		alias: {
-			$example: resolve('./src/example')
-		}
-	}
+	plugins: [sveltekit()]
 });
